fix: rename 'edit locally' to 'open locally'
Replace "Edit" by "Open"
